refactor(handoff): split complexity language heuristics
