+++ docs/architecture-consolidation-plan.md
@@ -67,7 +67,7 @@ fixtures:
 | Analysis preset grid | `crates/tokmd-analysis/src/grid/mod.rs`, `grid/presets.rs`, and `grid/disabled_feature.rs` | `mod.rs` 10; preset matrix owner 445; disabled-feature warning owner 66 | Keep `mod.rs` as the public barrel; keep preset identity, preset-to-enricher matrix, and plan lookup in `presets.rs`; keep disabled-feature warning text in `disabled_feature.rs`. Preserve the `analysis_orchestration` proof scope. |
 | Analysis API surface | `crates/tokmd-analysis/src/api_surface/mod.rs` and `crates/tokmd-analysis/src/api_surface/` | `mod.rs` 13; report owner 227; symbol dispatcher 56; language scanners <=68; symbol tests 385 | Keep `mod.rs` as a thin coordinator, report aggregation in `report.rs`, source scanning in language owner modules under `symbols`, and leave large integration tests under `api_surface/tests` |
 | Analysis effort implementation | `crates/tokmd-analysis/src/effort/size_basis.rs`, `crates/tokmd-analysis/src/effort/classify.rs`, and `crates/tokmd-analysis/src/effort/classify/` | size-basis aggregation 271; file classification coordinator 80; `.gitattributes` owner/tests 145; heuristic owner/tests 215 | Authored/generated/vendored line aggregation stays in `size_basis.rs`; `classify.rs` coordinates explicit rules and fallback classification; `.gitattributes` parsing and path-pattern matching live in `classify/gitattributes.rs`; generated sentinels and heuristic file tagging live in `classify/heuristics.rs`. Keep this under the `analysis_effort` proof scope. |
-| Context packing and handoff | `crates/tokmd/src/context_pack.rs`, `crates/tokmd/src/context_pack/`, `crates/tokmd/src/commands/context.rs`, and `crates/tokmd/src/commands/handoff.rs` | `context_pack.rs` 15; selection coordinator 305 + policy owner/tests 192 + selection tests 1368; ranking/packing owner 165; output/log submodule 224; render submodule 204; manifest submodule 195; budget parser/tests 220; context command 218; handoff command 268; handoff capability owner 113; handoff intelligence coordinator 196 + complexity owner 434; handoff output owner/tests 242 | Budget parsing, file selection, inclusion-policy preparation, selection proof tests, ranking/packing, bundle text rendering, single-output/log writing, context bundle manifest writing, handoff capability detection, handoff intelligence construction, lightweight source-complexity estimation, and handoff artifact output writing now live in owner modules; keep context/handoff proof scoped while future handoff splits depend on fresh evidence |
+| Context packing and handoff | `crates/tokmd/src/context_pack.rs`, `crates/tokmd/src/context_pack/`, `crates/tokmd/src/commands/context.rs`, and `crates/tokmd/src/commands/handoff.rs` | `context_pack.rs` 15; selection coordinator 305 + policy owner/tests 192 + selection tests 1368; ranking/packing owner 165; output/log submodule 224; render submodule 204; manifest submodule 195; budget parser/tests 220; context command 218; handoff command 268; handoff capability owner 113; handoff intelligence coordinator 196 + complexity coordinator 143 + language heuristic owner/tests 321; handoff output owner/tests 242 | Budget parsing, file selection, inclusion-policy preparation, selection proof tests, ranking/packing, bundle text rendering, single-output/log writing, context bundle manifest writing, handoff capability detection, handoff intelligence construction, lightweight source-complexity estimation, language-specific complexity heuristics, and handoff artifact output writing now live in owner modules; keep context/handoff proof scoped while future handoff splits depend on fresh evidence |
 | Format core outputs | `crates/tokmd-format/src/diff.rs`, `crates/tokmd-format/src/diff/`, `crates/tokmd-format/src/export.rs`, `crates/tokmd-format/src/export/`, `crates/tokmd-format/src/summary.rs`, and `crates/tokmd-format/src/summary/` | `diff.rs` 313; diff compute owner 168; diff render owner 274; export 101; export redaction owner 179; export CSV owner 47; export JSON owner 69; export JSONL owner 120; export CycloneDX owner 136; summary 424; summary JSON owner 117; language summary owner 86; module summary owner 53 | Diff row/total computation, diff Markdown rendering, export row redaction, CSV/JSON/JSONL/CycloneDX export rendering, language/module summary table rendering, and summary JSON receipt/file writing now live in dedicated owner modules while public helpers and output behavior stay unchanged; future summary coordinator reductions should be snapshot-backed and preserve the root `tokmd-format` re-exports |
 | Format fun outputs | `crates/tokmd-format/src/fun/mod.rs` and `crates/tokmd-format/src/fun/` | `mod.rs` 18; OBJ owner 136; MIDI owner 199 | OBJ code-city rendering and MIDI rendering now live in format-specific owner modules while the root fun module preserves public re-exports and feature-gated behavior |
 | Format redaction utilities | `crates/tokmd-format/src/redact/mod.rs` and `crates/tokmd-format/src/redact/` | `mod.rs` 251; safe extension owner 44 | Redaction hashing and path redaction remain in the public root module while safe file-extension preservation policy lives in its own private owner module; keep this under the `format_redaction_scan_args` proof scope |
